refactor(#779): pull out game table into own component
