+++ frontend/src/components/games/list/GameListItem.tsx
@@ -1,10 +1,10 @@
+import { GameInfo, GameResult } from '@/database/game';
 import { useLightMode } from '@/style/useLightMode';
 import CircleIcon from '@mui/icons-material/Circle';
 import CircleOutlinedIcon from '@mui/icons-material/CircleOutlined';
 import { Stack, Typography } from '@mui/material';
 import { GridRenderCellParams } from '@mui/x-data-grid-pro';
 import { useLocation } from 'react-router-dom';
-import { GameInfo, GameResult } from '../../database/game';
 
 interface RenderPlayersProps {
     white: string;
