+++ frontend/src/games/review/ReviewQueuePage.tsx
@@ -1,3 +1,11 @@
+import { useApi } from '@/api/Api';
+import { RequestSnackbar } from '@/api/Request';
+import { isValidDate } from '@/calendar/eventEditor/useEventEditor';
+import { RenderPlayersCell, RenderResult } from '@/components/games/list/GameListItem';
+import { CustomPagination } from '@/components/ui/CustomPagination';
+import { GameInfo, GameReviewType } from '@/database/game';
+import { usePagination } from '@/hooks/usePagination';
+import Avatar from '@/profile/Avatar';
 import { Container, Link, Stack, Typography } from '@mui/material';
 import {
     DataGridPro,
@@ -8,14 +16,6 @@ import {
 } from '@mui/x-data-grid-pro';
 import { useCallback } from 'react';
 import { Link as RouterLink, useNavigate } from 'react-router-dom';
-import { useApi } from '../../api/Api';
-import { RequestSnackbar } from '../../api/Request';
-import { isValidDate } from '../../calendar/eventEditor/useEventEditor';
-import { GameInfo, GameReviewType } from '../../database/game';
-import Avatar from '../../profile/Avatar';
-import { RenderPlayersCell, RenderResult } from '../list/GameListItem';
-import { CustomPagination } from '../list/ListGamesPage';
-import { usePagination } from '../list/pagination';
 
 export const ONE_WEEK = 1000 * 60 * 60 * 24 * 7;
