+++ frontend/src/profile/directories/AddExistingGamesDialog.tsx
@@ -1,9 +1,10 @@
 import { useApi } from '@/api/Api';
 import { RequestSnackbar, useRequest } from '@/api/Request';
 import { useRequiredAuth } from '@/auth/Auth';
+import { gameTableColumns } from '@/components/games/list/GameTable';
+import { CustomPagination } from '@/components/ui/CustomPagination';
 import { GameInfo } from '@/database/game';
-import { CustomPagination, gameTableColumns } from '@/games/list/ListGamesPage';
-import { usePagination } from '@/games/list/pagination';
+import { usePagination } from '@/hooks/usePagination';
 import { Directory } from '@jackstenglein/chess-dojo-common/src/database/directory';
 import { Visibility, VisibilityOff } from '@mui/icons-material';
 import { LoadingButton } from '@mui/lab';
