+++ frontend/src/profile/directories/DirectoryGridColumns.tsx
@@ -1,8 +1,11 @@
 import { useAuth } from '@/auth/Auth';
 import { toDojoDateString, toDojoTimeString } from '@/calendar/displayDate';
+import { RenderPlayers, RenderResult } from '@/components/games/list/GameListItem';
+import {
+    MastersCohort,
+    MastersOwnerDisplayName,
+} from '@/components/games/list/GameTable';
 import { dojoCohorts } from '@/database/user';
-import { RenderPlayers, RenderResult } from '@/games/list/GameListItem';
-import { MastersCohort, MastersOwnerDisplayName } from '@/games/list/ListGamesPage';
 import CohortIcon from '@/scoreboard/CohortIcon';
 import {
     DirectoryItem,
